requirements.txt
paramiko==3.3.1
fabric==2.7.1
ansible==8.5.0
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7
rich==13.6.0
tabulate==0.9.0 